Initial commit (not working)
Implementation as a composite action attempting to use caching with the
action/cache action as a nested action. However, this is currently not
possible with composite actions (see discussion in [1]). The ability
to invoke nested actions in composite actions will possibly be added in
the future.

So, there are two possibilities to implement this action: 1) keep using
a composite action but drop caching, 2) implement the action as a
JavaScript action, where it's possible to use caching with the @actions/cache
package of the JavaScript GitHub Actions toolkit ([2]).

[1] https://github.com/actions/runner/issues/646#issuecomment-704969347).
[2] https://github.com/actions/toolkit