Add optional 'version' input
The 'version' input allows to specify a specific version of the ajv-cli
package to install.

If this input is set, the action verifies whether the supplied version is
a valid version of ajv-cli and then installs this version.

If this input is not set, the action determines and installs the latest
version of the ajv-cli package.